fix fast refresh error by creating const files
