+++ viewer/src/constants/context.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const StoreContext = React.createContext(null);
